cloud-run/slack_app/src/requirements.txt
functions-framework==2.0.0
slack_sdk>=3.0
slackeventsapi>=2.1.0
Flask>=1.1.2
google-cloud-firestore==1.9.0
requests